Add initial HTML structure for Christmas raffle
I need you to generate a static web page for me that I can deploy to Github Pages. I want to use this web page to help with our Christmas present raffle. The idea is that each person in our group draws one other person from the group. They would then get a present for only their assigned person. This way, everyone gets a present at the end, and everyone only gives one present (and they know who are giving it to, but the person who gets it won't necessarily know who it is from).

I want to use the web page to hold a static assignment of the randomly initialized assignment. I was thinking that the web page should have one input field to enter the first name, and if such a name exists in the pool, would present the name of the person they will give a present to. It's OK if the allocation is easily visible from the source code - I don't expect anyone in our group to actually look at the source code of the page (this would spoil the fun for everyone). However, to make it at least a little less obvious, you could first convert any entered name into lowercase and then calculate a hash for that name (e.g. MD5), and then compare that to pre-initialized hashes to return the plain name of the associated person. This way even just looking at the source code wouldn't make it obvious who this person is associated with.

Again, it is very important that everyone has exactly one person assigned, and every person is assigned exactly once (so that everyone gets a present).

The names of our group are: Enrico, Nadine, Sarah, Ricci, Diana

A few more notes:
* Make sure to use the lowercase names to calculate the MD5 hash for the comparison, as you will trim and lowercase the names in the actual web page, too
* Do not include the plain text assignments anywhere in the code - not even comments. The purpose of using the MD5 hash is so that someone looking at the source code does not immediately see the assignment

Alright, I think that's all you need to know. Do you have any questions? If not, please generate that static web page. Thakns